requirements.txt
aioredis==1.3.1
asgiref==3.5.2
async-timeout==4.0.2
autobahn==22.7.1
Automat==20.2.0
Brotli==1.0.9
cachetools==5.2.0
channels==2.4.0
channels-redis==3.4.1
charset-normalizer==2.1.0
click==8.1.3
constantly==15.1.0
cryptography==38.0.1
daphne==2.5.0
dash==2.6.1
dash-bootstrap-components==0.13.1
dash-core-components==2.0.0
dash-daq==0.5.0
dash-html-components==2.0.0
dash-table==5.0.0
Deprecated==1.2.13
Django==3.2.15
django-plotly-dash==2.0.0
django-redis==5.2.0
dpd-components==0.1.0
Flask==2.1.3
Flask-Compress==1.12
fonttools==4.25.0
google-api-core==2.8.2
google-api-python-client==2.56.0
google-auth==2.10.0
google-auth-httplib2==0.1.0
googleapis-common-protos==1.56.4
gunicorn==20.1.0
hiredis==2.0.0
httplib2==0.20.4
hyperlink==21.0.0
idna==3.3
importlib-metadata==4.12.0
incremental==21.3.0
itsdangerous==2.1.2
mkl-fft==1.3.1
mkl-service==2.4.0
msgpack==1.0.4
munkres==1.1.4
Pillow==9.2.0
plotly==5.10.0
ply==3.11
protobuf==4.21.5
psycopg2==2.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyOpenSSL==22.0.0
pyparsing==3.0.9
PyQt5==5.15.7
redis==4.3.4
requests==2.28.1
rsa==4.9
service-identity==21.1.0
sqlparse==0.4.2
tenacity==8.0.1
Twisted==22.8.0
txaio==22.2.1
tzdata==2022.2
uritemplate==4.1.1
urllib3==1.26.11
webencodings==0.5.1
Werkzeug==2.0.3
wincertstore==0.2
wrapt==1.14.1
zipp==3.8.1
zope.interface==5.4.0
whitenoise==6.2.0
python-decouple==3.6